refactor: move MessageDetailsResponse
