+++ Doppler.PushContact.Models/PushContactApiResponses/MessageDetailsResponse.cs
@@ -1,6 +1,4 @@
-using System;
-
-namespace Doppler.PushContact.ApiModels
+namespace Doppler.PushContact.Models.PushContactApiResponses
 {
     public class MessageDetailsResponse
     {
